added elf_version_definitions table
Added the elf_version_definitions table which correlates to
.gnu.version_d.

Not 100% sure this table is needed since we capture the version
information in the symbol table as well, but this is a closer
approximation to the ELF format itself which should help others
navigate.

It would be great to have a foreign key reference here as well perhaps i
the future.